refactor(components): detach atom component for modal
